requirements.txt
Django>=3.0,<4.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
django-markdownx